server/requirements.txt
fastapi
uvicorn
motor
redis
python-dotenv
apscheduler
pydantic[email]
pytz
croniter
celery
celery[redis]
